refactor: delete block button, use fab
