+++ apps/journeys-admin/src/components/Editor/AddBlockDrawer/AddBlockDrawer.tsx
@@ -2,15 +2,15 @@ import { ReactElement } from 'react'
 
 import Stack from '@mui/material/Stack'
 
-import { NewTypographyButton } from '../Canvas/AddBlockButton/NewTypographyButton'
-import { NewImageButton } from '../Canvas/AddBlockButton/NewImageButton'
-import { NewVideoButton } from '../Canvas/AddBlockButton/NewVideoButton'
-import { NewRadioQuestionButton } from '../Canvas/AddBlockButton/NewRadioQuestionButton'
-import { NewTextResponseButton } from '../Canvas/AddBlockButton/NewTextResponseButton'
-import { NewSignUpButton } from '../Canvas/AddBlockButton/NewSignUpButton'
-import { NewButtonButton } from '../Canvas/AddBlockButton/NewButtonButton'
+import { NewTypographyButton } from '../Canvas/Button/NewTypographyButton'
+import { NewImageButton } from '../Canvas/Button/NewImageButton'
+import { NewVideoButton } from '../Canvas/Button/NewVideoButton'
+import { NewRadioQuestionButton } from '../Canvas/Button/NewRadioQuestionButton'
+import { NewTextResponseButton } from '../Canvas/Button/NewTextResponseButton'
+import { NewSignUpButton } from '../Canvas/Button/NewSignUpButton'
+import { NewButtonButton } from '../Canvas/Button/NewButtonButton'
 import { useFlags } from '@core/shared/ui/FlagsProvider'
-import { NewFormButton } from '../Canvas/AddBlockButton/NewFormButton'
+import { NewFormButton } from '../Canvas/Button/NewFormButton'
 import { Grid, useMediaQuery } from '@mui/material'
 import { Theme } from '@mui/material/styles'
