--- apps/journeys-admin/src/components/Editor/Canvas/AddBlockButton/index.ts
@@ -1,1 +1,0 @@
-export { AddBlockButton } from './AddBlockButton'
