+++ apps/journeys-admin/src/components/Editor/Canvas/Button/NewImageButton/NewImageButton.spec.tsx
