+++ apps/journeys-admin/src/components/Editor/Canvas/Button/NewImageButton/index.ts
