+++ apps/journeys-admin/src/components/Editor/Canvas/Button/NewSignUpButton/NewSignUpButton.spec.tsx
